+++ doc/conf.py
@@ -34,6 +34,7 @@ extensions = '''
     sphinx.ext.todo
     sphinx.ext.viewcode
     sphinx.ext.autosummary
+    sphinx_rtd_theme
 '''.split()
 
 # Add any paths that contain templates here, relative to this directory.
@@ -103,7 +104,7 @@ pygments_style = 'sphinx'
 # html_theme = 'nature'
 # Use default rather than my normal nature so we get the read-the-docs style on
 # that website.
-html_theme = 'default'
+html_theme = 'sphinx_rtd_theme'
 
 # Theme options are theme-specific and customize the look and feel of a theme
 # further.  For a list of options available for each theme, see the